Crate swc_ecma_transforms

Re-exports

pub use swc_ecma_transforms_compat as compat;

pub use swc_ecma_transforms_module as modules;

pub use swc_ecma_transforms_optimization as optimization;

pub use swc_ecma_transforms_proposal as proposals;

pub use swc_ecma_transforms_react as react;

pub use swc_ecma_transforms_typescript as typescript;

Modules

feature

fixer

helpers

hygiene

pass

perf

Structs

Assumptions

Alternative for https://babeljs.io/docs/en/assumptions

Functions

const_modules

fixer

Fixes ast nodes before printing so semantics are preserved.

hygiene

See hygiene_with_config for doc. Creates a hygiene pass with default value of Config.

resolver

See Ident for know how does swc manages identifiers.