swc_ecma_transforms::perf
Trait ParallelExt

pub trait ParallelExt: Parallel {
    // Required method
    fn maybe_par_idx<I, F>(&mut self, threshold: usize, nodes: I, op: F)
       where I: Items,
             F: Send + Sync + Fn(&mut Self, usize, <I as Items>::Elem);

    // Provided method
    fn maybe_par<I, F>(&mut self, threshold: usize, nodes: I, op: F)
       where I: Items,
             F: Send + Sync + Fn(&mut Self, <I as Items>::Elem) { ... }
}

Required Methods

fn maybe_par_idx<I, F>(&mut self, threshold: usize, nodes: I, op: F)
where I: Items, F: Send + Sync + Fn(&mut Self, usize, <I as Items>::Elem),

Invoke op in parallel, if swc_ecma_utils is compiled with concurrent feature enabled and nodes.len() is bigger than threshold.

This configures GLOBALS, while not configuring [HANDLER] nor [HELPERS]

Provided Methods

fn maybe_par<I, F>(&mut self, threshold: usize, nodes: I, op: F)
where I: Items, F: Send + Sync + Fn(&mut Self, <I as Items>::Elem),

Invoke op in parallel, if swc_ecma_utils is compiled with concurrent feature enabled and nodes.len() is bigger than threshold.

This configures GLOBALS, while not configuring [HANDLER] nor [HELPERS]

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.