Struct swc_ecma_transforms::feature::FeatureFlag

pub struct FeatureFlag { /* private fields */ }

Implementations

impl FeatureFlag

pub const TemplateLiterals: FeatureFlag = Self{ bits: 1 << 0,}

transform-template-literals

pub const Literals: FeatureFlag = Self{ bits: 1 << 1,}

transform-literals

pub const FunctionName: FeatureFlag = Self{ bits: 1 << 2,}

transform-function-name

pub const ArrowFunctions: FeatureFlag = Self{ bits: 1 << 3,}

transform-arrow-functions

pub const BlockScopedFunctions: FeatureFlag = Self{ bits: 1 << 4,}

transform-block-scoped-functions

pub const Classes: FeatureFlag = Self{ bits: 1 << 5,}

transform-classes

pub const ObjectSuper: FeatureFlag = Self{ bits: 1 << 6,}

transform-object-super

pub const ShorthandProperties: FeatureFlag = Self{ bits: 1 << 7,}

transform-shorthand-properties

pub const DuplicateKeys: FeatureFlag = Self{ bits: 1 << 8,}

transform-duplicate-keys

pub const ComputedProperties: FeatureFlag = Self{ bits: 1 << 9,}

transform-computed-properties

pub const ForOf: FeatureFlag = Self{ bits: 1 << 10,}

transform-for-of

pub const StickyRegex: FeatureFlag = Self{ bits: 1 << 11,}

transform-sticky-regex

pub const DotAllRegex: FeatureFlag = Self{ bits: 1 << 12,}

transform-dotall-regex

pub const UnicodeRegex: FeatureFlag = Self{ bits: 1 << 13,}

transform-unicode-regex

pub const Spread: FeatureFlag = Self{ bits: 1 << 14,}

transform-spread

pub const Parameters: FeatureFlag = Self{ bits: 1 << 15,}

transform-parameters

pub const Destructuring: FeatureFlag = Self{ bits: 1 << 16,}

transform-destructuring

pub const BlockScoping: FeatureFlag = Self{ bits: 1 << 17,}

transform-block-scoping

pub const TypeOfSymbol: FeatureFlag = Self{ bits: 1 << 18,}

transform-typeof-symbol

pub const NewTarget: FeatureFlag = Self{ bits: 1 << 19,}

transform-new-target

pub const Regenerator: FeatureFlag = Self{ bits: 1 << 20,}

transform-regenerator

pub const ExponentiationOperator: FeatureFlag = Self{ bits: 1 << 21,}

transform-exponentiation-operator

pub const AsyncToGenerator: FeatureFlag = Self{ bits: 1 << 22,}

transform-async-to-generator

pub const AsyncGeneratorFunctions: FeatureFlag = Self{ bits: 1 << 23,}

proposal-async-generator-functions

pub const ObjectRestSpread: FeatureFlag = Self{ bits: 1 << 24,}

proposal-object-rest-spread

pub const UnicodePropertyRegex: FeatureFlag = Self{ bits: 1 << 25,}

proposal-unicode-property-regex

pub const JsonStrings: FeatureFlag = Self{ bits: 1 << 26,}

proposal-json-strings

pub const OptionalCatchBinding: FeatureFlag = Self{ bits: 1 << 27,}

proposal-optional-catch-binding

pub const NamedCapturingGroupsRegex: FeatureFlag = Self{ bits: 1 << 28,}

transform-named-capturing-groups-regex

pub const MemberExpressionLiterals: FeatureFlag = Self{ bits: 1 << 29,}

transform-member-expression-literals

pub const PropertyLiterals: FeatureFlag = Self{ bits: 1 << 30,}

transform-property-literals

pub const ReservedWords: FeatureFlag = Self{ bits: 1 << 31,}

transform-reserved-words

pub const ExportNamespaceFrom: FeatureFlag = Self{ bits: 1 << 32,}

proposal-export-namespace-from

pub const NullishCoalescing: FeatureFlag = Self{ bits: 1 << 33,}

proposal-nullish-coalescing-operator

pub const LogicalAssignmentOperators: FeatureFlag = Self{ bits: 1 << 34,}

proposal-logical-assignment-operators

pub const OptionalChaining: FeatureFlag = Self{ bits: 1 << 35,}

proposal-optional-chaining

pub const ClassProperties: FeatureFlag = Self{ bits: 1 << 36,}

proposal-class-properties

pub const NumericSeparator: FeatureFlag = Self{ bits: 1 << 37,}

proposal-numeric-separator

pub const PrivateMethods: FeatureFlag = Self{ bits: 1 << 38,}

proposal-private-methods

pub const ClassStaticBlock: FeatureFlag = Self{ bits: 1 << 39,}

proposal-class-static-block

pub const PrivatePropertyInObject: FeatureFlag = Self{ bits: 1 << 40,}

proposal-private-property-in-object

pub const UnicodeEscapes: FeatureFlag = Self{ bits: 1 << 41,}

transform-unicode-escapes

pub const BugfixAsyncArrowsInClass: FeatureFlag = Self{ bits: 1 << 42,}

bugfix/transform-async-arrows-in-class

pub const BugfixEdgeDefaultParam: FeatureFlag = Self{ bits: 1 << 43,}

bugfix/transform-edge-default-parameters

pub const BugfixTaggedTemplateCaching: FeatureFlag = Self{ bits: 1 << 44,}

bugfix/transform-tagged-template-caching

pub const BugfixSafariIdDestructuringCollisionInFunctionExpression: FeatureFlag = Self{ bits: 1 << 45,}

bugfix/transform-safari-id-destructuring-collision-in-function-expression

pub const BugfixTransformEdgeFunctionName: FeatureFlag = Self{ bits: 1 << 46,}

bugfix/transform-edge-function-name

pub const BugfixTransformSafariBlockShadowing: FeatureFlag = Self{ bits: 1 << 47,}

bugfix/transform-safari-block-shadowing

pub const BugfixTransformSafariForShadowing: FeatureFlag = Self{ bits: 1 << 48,}

bugfix/transform-safari-for-shadowing

pub const BugfixTransformV8SpreadParametersInOptionalChaining: FeatureFlag = Self{ bits: 1 << 49,}

bugfix/transform-v8-spread-parameters-in-optional-chaining

pub const fn empty() -> FeatureFlag

Returns an empty set of flags.

pub const fn all() -> FeatureFlag

Returns the set containing all flags.

pub const fn bits(&self) -> u64

Returns the raw value of the flags currently stored.

pub const fn from_bits(bits: u64) -> Option<FeatureFlag>

Convert from underlying bit representation, unless that representation contains bits that do not correspond to a flag.

pub const fn from_bits_truncate(bits: u64) -> FeatureFlag

Convert from underlying bit representation, dropping any bits that do not correspond to flags.

pub const unsafe fn from_bits_unchecked(bits: u64) -> FeatureFlag

Convert from underlying bit representation, preserving all bits (even those not corresponding to a defined flag).

Safety

The caller of the bitflags! macro can chose to allow or disallow extra bits for their bitflags type.

The caller of from_bits_unchecked() has to ensure that all bits correspond to a defined flag or that extra bits are valid for this bitflags type.

pub const fn is_empty(&self) -> bool

Returns true if no flags are currently stored.

pub const fn is_all(&self) -> bool

Returns true if all flags are currently set.

pub const fn intersects(&self, other: FeatureFlag) -> bool

Returns true if there are flags common to both self and other.

pub const fn contains(&self, other: FeatureFlag) -> bool

Returns true if all of the flags in other are contained within self.

pub fn insert(&mut self, other: FeatureFlag)

Inserts the specified flags in-place.

pub fn remove(&mut self, other: FeatureFlag)

Removes the specified flags in-place.

pub fn toggle(&mut self, other: FeatureFlag)

Toggles the specified flags in-place.

pub fn set(&mut self, other: FeatureFlag, value: bool)

Inserts or removes the specified flags depending on the passed value.

pub const fn intersection(self, other: FeatureFlag) -> FeatureFlag

Returns the intersection between the flags in self and other.

Specifically, the returned set contains only the flags which are present in both self and other.

This is equivalent to using the & operator (e.g. ops::BitAnd), as in flags & other.

pub const fn union(self, other: FeatureFlag) -> FeatureFlag

Returns the union of between the flags in self and other.

Specifically, the returned set contains all flags which are present in either self or other, including any which are present in both (see Self::symmetric_difference if that is undesirable).

This is equivalent to using the | operator (e.g. ops::BitOr), as in flags | other.

pub const fn difference(self, other: FeatureFlag) -> FeatureFlag

Returns the difference between the flags in self and other.

Specifically, the returned set contains all flags present in self, except for the ones present in other.

It is also conceptually equivalent to the “bit-clear” operation: flags & !other (and this syntax is also supported).

This is equivalent to using the - operator (e.g. ops::Sub), as in flags - other.

pub const fn symmetric_difference(self, other: FeatureFlag) -> FeatureFlag

Returns the symmetric difference between the flags in self and other.

Specifically, the returned set contains the flags present which are present in self or other, but that are not present in both. Equivalently, it contains the flags present in exactly one of the sets self and other.

This is equivalent to using the ^ operator (e.g. ops::BitXor), as in flags ^ other.

pub const fn complement(self) -> FeatureFlag

Returns the complement of this set of flags.

Specifically, the returned set contains all the flags which are not set in self, but which are allowed for this type.

Alternatively, it can be thought of as the set difference between Self::all() and self (e.g. Self::all() - self)

This is equivalent to using the ! operator (e.g. ops::Not), as in !flags.